+++ lib/core/extension/scale_down_box.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+extension ScaleDownBox on Text {
+  FittedBox get scaleDown => FittedBox(
+        fit: BoxFit.scaleDown,
+        child: this,
+      );
+}