+++ lib/feature/feature_dialog_box/date_picker_dialog_box_without_day.dart
@@ -1,6 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:persian_calendar_widget/core/enum/enum.dart';
 import 'package:persian_calendar_widget/core/extension/space_xy.dart';
+import 'package:persian_calendar_widget/core/utils/constants/app_constants.dart';
+import 'package:persian_calendar_widget/core/widgets/cancel_button.dart';
+import 'package:persian_calendar_widget/core/widgets/choose_button.dart';
+import 'package:persian_calendar_widget/core/widgets/month_page_view.dart';
+import 'package:persian_calendar_widget/core/widgets/today_context.dart';
+import 'package:persian_calendar_widget/core/widgets/type_data_button.dart';
+import 'package:persian_calendar_widget/core/widgets/year_page_view.dart';
 import 'package:persian_calendar_widget/persian_calendar_widget.dart';
 import 'package:shamsi_date/shamsi_date.dart';
 
@@ -39,7 +46,6 @@ class DatePickerDialogBoxWithoutDay extends StatefulWidget {
   final ButtonStyle? goButtonStyle;
   final TextStyle? goTextStyle;
   final bool showTodayBanner;
-  final TextStyle? todayTitleBannerTextStyle;
   final TextStyle? todayDateBannerTextStyle;
 
   const DatePickerDialogBoxWithoutDay({
@@ -73,7 +79,6 @@ class DatePickerDialogBoxWithoutDay extends StatefulWidget {
     required this.goTitle,
     required this.showTodayBanner,
     required this.todayDateBannerTextStyle,
-    required this.todayTitleBannerTextStyle,
     super.key,
   });
 
@@ -100,21 +105,6 @@ class _DatePickerDialogBoxWithoutDayState
   // late int selectedDay;
   late int selectedMonths;
   late int selectedYear;
-
-  final Map<int, String> monthsMap = {
-    1: 'فروردین',
-    2: 'اردیبهشت',
-    3: 'خرداد',
-    4: 'تیر',
-    5: 'مرداد',
-    6: 'شهریور',
-    7: 'مهر',
-    8: 'آبان',
-    9: 'آذر',
-    10: 'دی',
-    11: 'بهمن',
-    12: 'اسفند',
-  };
 
   @override
   void initState() {
@@ -173,8 +163,10 @@ class _DatePickerDialogBoxWithoutDayState
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: [
+                12.0.spaceX,
+
                 /// month button
-                TextButton(
+                TypeDataButton(
                   onPressed: () {
                     _pageController.animateToPage(
                       PageViewIndex.month,
@@ -185,36 +177,19 @@ class _DatePickerDialogBoxWithoutDayState
                       currentPageViewIndex = PageViewIndex.month;
                     });
                   },
-                  style: widget.titleButtonStyle == null
-                      ? TextButton.styleFrom(
-                          backgroundColor:
-                              currentPageViewIndex == PageViewIndex.month
-                                  ? widget.onPrimaryColor ??
-                                      Theme.of(context).scaffoldBackgroundColor
-                                  : Colors.transparent,
-                        )
-                      : currentPageViewIndex == PageViewIndex.month
-                          ? widget.titleSelectedButtonStyle ??
-                              widget.titleButtonStyle
-                          : widget.titleButtonStyle,
-                  child: Text(
-                    formattedDate.mN,
-                    style: widget.titleTextStyle == null
-                        ? TextStyle(
-                            color: currentPageViewIndex == PageViewIndex.month
-                                ? widget.primaryColor ??
-                                    Theme.of(context).primaryColor
-                                : widget.onPrimaryColor ?? Colors.white,
-                          )
-                        : currentPageViewIndex == PageViewIndex.month
-                            ? widget.titleSelectedTextStyle ??
-                                widget.titleTextStyle
-                            : widget.titleTextStyle,
-                  ),
+                  buttonStyle: widget.titleButtonStyle,
+                  selectedButtonStyle: widget.titleSelectedButtonStyle,
+                  title: formattedDate.mN,
+                  isSelected: currentPageViewIndex == PageViewIndex.month,
+                  textStyle: widget.titleTextStyle,
+                  selectedTextStyle: widget.titleSelectedTextStyle,
+                  onPrimaryColor: widget.onPrimaryColor,
+                  primaryColor: widget.primaryColor,
                 ),
+                5.0.spaceX,
 
                 /// year button
-                TextButton(
+                TypeDataButton(
                   onPressed: () {
                     _pageController.animateToPage(
                       PageViewIndex.year,
@@ -225,141 +200,40 @@ class _DatePickerDialogBoxWithoutDayState
                       currentPageViewIndex = PageViewIndex.year;
                     });
                   },
-                  style: widget.titleButtonStyle == null
-                      ? TextButton.styleFrom(
-                          backgroundColor:
-                              currentPageViewIndex == PageViewIndex.year
-                                  ? widget.onPrimaryColor ??
-                                      Theme.of(context).scaffoldBackgroundColor
-                                  : Colors.transparent,
-                        )
-                      : currentPageViewIndex == PageViewIndex.year
-                          ? widget.titleSelectedButtonStyle ??
-                              widget.titleButtonStyle
-                          : widget.titleButtonStyle,
-                  child: Text(
-                    widget.calendarType == CalendarType.persian
-                        ? formattedDate.yyyy.toPersianDigit()
-                        : formattedDate.yyyy,
-                    style: widget.titleTextStyle == null
-                        ? TextStyle(
-                            color: currentPageViewIndex == PageViewIndex.year
-                                ? widget.primaryColor ??
-                                    Theme.of(context).primaryColor
-                                : widget.onPrimaryColor ?? Colors.white,
-                          )
-                        : currentPageViewIndex == PageViewIndex.year
-                            ? widget.titleSelectedTextStyle ??
-                                widget.titleTextStyle
-                            : widget.titleTextStyle,
-                  ),
+                  buttonStyle: widget.titleButtonStyle,
+                  selectedButtonStyle: widget.titleSelectedButtonStyle,
+                  title: widget.calendarType == CalendarType.persian
+                      ? formattedDate.yyyy.toPersianDigit()
+                      : formattedDate.yyyy,
+                  isSelected: currentPageViewIndex == PageViewIndex.year,
+                  textStyle: widget.titleTextStyle,
+                  selectedTextStyle: widget.titleSelectedTextStyle,
+                  onPrimaryColor: widget.onPrimaryColor,
+                  primaryColor: widget.primaryColor,
                 ),
+                12.0.spaceX,
               ],
             ),
           ),
-          if (widget.showTodayBanner || widget.useGoToTodayButton) 3.spaceY,
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 12),
-            child: Row(
-              children: [
-                if (widget.useGoToTodayButton) ...[
-                  1.spaceX,
-                  Expanded(
-                    child: SizedBox(
-                      height: 27,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          _updateSelectedDate(today);
-                          setState(() {});
-                        },
-                        style: widget.goButtonStyle ??
-                            ElevatedButton.styleFrom(
-                              padding: EdgeInsets.zero,
-                              backgroundColor: widget.onPrimaryColor,
-                              shape: RoundedRectangleBorder(
-                                /// elevated button border radius fix by main
-                                /// box border radius value
-                                borderRadius: widget.showTodayBanner
-                                    ? BorderRadius.only(
-                                        topLeft: const Radius.circular(5),
-                                        topRight: const Radius.circular(5),
-                                        bottomLeft: const Radius.circular(5),
-                                        bottomRight: Radius.circular(
-                                          widget.borderRadius * 0.6,
-                                        ),
-                                      )
-                                    : BorderRadius.vertical(
-                                        bottom: Radius.circular(
-                                          widget.borderRadius * .6,
-                                        ),
-                                        top: const Radius.circular(5),
-                                      ),
-                              ),
-                            ),
-                        child: Text(
-                          widget.goTitle ??
-                              (widget.showTodayBanner
-                                  ? 'برو به'
-                                  : 'برو به امروز'),
-                          style: widget.goTextStyle ??
-                              TextStyle(
-                                color: widget.primaryColor,
-                              ),
-                        ),
-                      ),
-                    ),
-                  ),
-                  1.spaceX,
-                ],
-                if (widget.showTodayBanner && widget.useGoToTodayButton)
-                  3.spaceX,
-                if (widget.showTodayBanner)
-                  Expanded(
-                    flex: 2,
-                    child: DecoratedBox(
-                      decoration: BoxDecoration(
-                        borderRadius: widget.useGoToTodayButton
-                            ? BorderRadius.only(
-                                topLeft: const Radius.circular(5),
-                                topRight: const Radius.circular(5),
-                                bottomRight: const Radius.circular(5),
-                                bottomLeft:
-                                    Radius.circular(widget.borderRadius * 0.6),
-                              )
-                            : BorderRadius.vertical(
-                                bottom:
-                                    Radius.circular(widget.borderRadius * .6),
-                                top: const Radius.circular(5),
-                              ),
-                        border: Border.all(
-                          color: widget.primaryColor ??
-                              Theme.of(context).primaryColor,
-                          width: .5,
-                        ),
-                      ),
-                      child: Padding(
-                        padding: const EdgeInsets.symmetric(vertical: 3),
-                        child: Text(
-                          widget.calendarType == CalendarType.persian
-                              ?
-                              // ignore: lines_longer_than_80_chars
-                              ' ${today.formatter.wN} ${today.day} ${today.formatter.mN} ${today.year}'
-                                  .toPersianDigit()
-                              // ignore: lines_longer_than_80_chars
-                              : ' ${today.formatter.wN} ${today.day} ${today.formatter.mN} ${today.year}',
-                          style: widget.todayDateBannerTextStyle ??
-                              TextStyle(
-                                color: widget.primaryColor ??
-                                    Theme.of(context).primaryColor,
-                              ),
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                    ),
-                  ),
-              ],
+          if (widget.showTodayBanner || widget.useGoToTodayButton) ...[
+            3.spaceY,
+            TodayContext(
+              useGoToTodayButton: widget.useGoToTodayButton,
+              showTodayBanner: widget.showTodayBanner,
+              onPressed: (today) {
+                _updateSelectedDate(today);
+                setState(() {});
+              },
+              buttonStyle: widget.goButtonStyle,
+              onPrimaryColor: widget.onPrimaryColor,
+              primaryColor: widget.primaryColor,
+              borderRadius: widget.borderRadius,
+              isPersian: widget.calendarType == CalendarType.persian,
+              title: widget.goTitle,
+              goTextStyle: widget.goTextStyle,
+              todayDateBannerTextStyle: widget.todayDateBannerTextStyle,
             ),
-          ),
+          ],
 
           SizedBox(
             width: MediaQuery.sizeOf(context).width,
@@ -369,121 +243,42 @@ class _DatePickerDialogBoxWithoutDayState
               physics: const NeverScrollableScrollPhysics(),
               children: [
                 /// month screen
-                GridView.builder(
-                  padding: const EdgeInsets.only(
-                    right: 12,
-                    left: 12,
-                    top: 10,
-                    bottom: 5,
-                  ),
-                  shrinkWrap: true,
-                  gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
-                    maxCrossAxisExtent: 100,
-                    childAspectRatio: 1.7,
-                  ),
-                  itemCount: monthsMap.length,
-                  itemBuilder: (context, index) {
-                    final int currentMonth = index + 1;
-
-                    return TextButton(
-                      style: widget.dateButtonStyle == null
-                          ? TextButton.styleFrom(
-                              backgroundColor: selectedMonths == currentMonth
-                                  ? widget.primaryColor ??
-                                      Theme.of(context).primaryColor
-                                  : Colors.transparent,
-                              maximumSize: const Size.fromWidth(100),
-                              minimumSize: const Size.fromWidth(100),
-                              padding: EdgeInsets.zero,
-
-                              // fixedSize: const Size(100, 0),
-                            )
-                          : selectedMonths == currentMonth
-                              ? widget.dateSelectedButtonStyle ??
-                                  widget.dateButtonStyle
-                              : widget.dateButtonStyle,
-                      child: Text(
-                        '${monthsMap[currentMonth]}',
-                        style: widget.dateTextStyle == null
-                            ? TextStyle(
-                                color: selectedMonths == currentMonth
-                                    ? widget.onPrimaryColor ?? Colors.white
-                                    : widget.primaryColor,
-                              )
-                            : selectedMonths == currentMonth
-                                ? widget.dateSelectedTextStyle ??
-                                    widget.dateTextStyle
-                                : widget.dateTextStyle,
-                      ),
-                      onPressed: () {
-                        /// set new date with new index
-                        _updateSelectedDate(
-                          selectedDate.withMonth(currentMonth),
-                        );
-                        setState(() {});
-                      },
+                MonthPageView(
+                  dateButtonStyle: widget.dateButtonStyle,
+                  selectedMonths: selectedMonths,
+                  primaryColor: widget.primaryColor,
+                  onPrimaryColor: widget.onPrimaryColor,
+                  dateSelectedTextStyle: widget.dateSelectedTextStyle,
+                  dateTextStyle: widget.dateTextStyle,
+                  dateSelectedButtonStyle: widget.dateSelectedButtonStyle,
+                  onPressed: (currentMonth) {
+                    /// set new date with new index
+                    _updateSelectedDate(
+                      selectedDate.withMonth(currentMonth),
                     );
+                    setState(() {});
                   },
                 ),
 
                 /// year screen
-                GridView.builder(
-                  padding: const EdgeInsets.only(
-                    right: 12,
-                    left: 12,
-                    top: 10,
-                    bottom: 5,
-                  ),
-                  shrinkWrap: true,
-                  gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
-                    maxCrossAxisExtent: 100,
-                    childAspectRatio: 1.7,
-                  ),
-                  itemCount: yearsList.length,
-                  itemBuilder: (context, index) {
-                    final int currentYear = index;
-
-                    return TextButton(
-                      style: widget.dateButtonStyle == null
-                          ? TextButton.styleFrom(
-                              backgroundColor:
-                                  selectedYear == currentYear + minYear
-                                      ? widget.primaryColor ??
-                                          Theme.of(context).primaryColor
-                                      : Colors.transparent,
-                              maximumSize: const Size.fromWidth(100),
-                              minimumSize: const Size.fromWidth(100),
-                              padding: EdgeInsets.zero,
-
-                              // fixedSize: const Size(100, 0),
-                            )
-                          : selectedYear == currentYear + minYear
-                              ? widget.dateSelectedButtonStyle ??
-                                  widget.dateButtonStyle
-                              : widget.dateButtonStyle,
-                      child: Text(
-                        widget.calendarType == CalendarType.persian
-                            ? '${yearsList[index]}'.toPersianDigit()
-                            : '${yearsList[index]}',
-                        style: widget.dateTextStyle == null
-                            ? TextStyle(
-                                color: selectedYear == currentYear + minYear
-                                    ? widget.onPrimaryColor ?? Colors.white
-                                    : widget.primaryColor,
-                              )
-                            : selectedYear == currentYear + minYear
-                                ? widget.dateSelectedTextStyle ??
-                                    widget.dateTextStyle
-                                : widget.dateTextStyle,
-                      ),
-                      onPressed: () {
-                        /// set new date with new index
-                        _updateSelectedDate(
-                          selectedDate.withYear(currentYear + minYear),
-                        );
-                        setState(() {});
-                      },
+                YearPageView(
+                  listLength: yearsList.length,
+                  dateButtonStyle: widget.dateButtonStyle,
+                  selectedYear: selectedYear,
+                  minYear: minYear,
+                  primaryColor: widget.primaryColor,
+                  onPrimaryColor: widget.onPrimaryColor,
+                  isPersian: widget.calendarType == CalendarType.persian,
+                  yearsList: yearsList,
+                  dateSelectedTextStyle: widget.dateSelectedTextStyle,
+                  dateTextStyle: widget.dateTextStyle,
+                  dateSelectedButtonStyle: widget.dateSelectedButtonStyle,
+                  onPressed: (currentYear) {
+                    /// set new date with new index
+                    _updateSelectedDate(
+                      selectedDate.withYear(currentYear + minYear),
                     );
+                    setState(() {});
                   },
                 ),
               ],
@@ -493,59 +288,30 @@ class _DatePickerDialogBoxWithoutDayState
             padding: const EdgeInsets.symmetric(horizontal: 18),
             child: Row(
               children: [
-                Expanded(
-                  child: ElevatedButton(
-                    onPressed: widget.onSubmit != null
-                        ? () {
-                            widget.onSubmit!(
-                              selectedDate,
-                              selectedDateInText,
-                            );
-                            Navigator.pop(context);
-                          }
-                        : null,
-                    style: widget.submitButtonStyle ??
-                        ElevatedButton.styleFrom(
-                          backgroundColor: widget.onPrimaryColor,
-                          shape: RoundedRectangleBorder(
-                            /// elevated button border radius fix by main box
-                            /// border radius value
-                            borderRadius: BorderRadius.circular(
-                              widget.borderRadius * 0.6,
-                            ),
-                          ),
-                        ),
-                    child: Text(
-                      widget.submitTitle ?? 'انتخاب',
-                      style: widget.submitTextStyle ??
-                          TextStyle(
-                            color: widget.primaryColor,
-                          ),
-                    ),
-                  ),
+                ChooseButton(
+                  onPressed: widget.onSubmit != null
+                      ? () {
+                          widget.onSubmit!(
+                            selectedDate,
+                            selectedDateInText,
+                          );
+                          Navigator.pop(context);
+                        }
+                      : null,
+                  buttonStyle: widget.submitButtonStyle,
+                  borderRadius: widget.borderRadius,
+                  title: widget.submitTitle,
+                  textStyle: widget.submitTextStyle,
+                  primaryColor: widget.primaryColor,
+                  onPrimaryColor: widget.onPrimaryColor,
                 ),
-                const SizedBox(
-                  width: 8,
-                ),
-                TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  style: widget.cancelButtonStyle ??
-                      ElevatedButton.styleFrom(
-                        shape: RoundedRectangleBorder(
-                          /// text button border radius fix by main box border
-                          /// radius value
-                          borderRadius: BorderRadius.circular(
-                            widget.borderRadius * 0.6,
-                          ),
-                        ),
-                      ),
-                  child: Text(
-                    widget.cancelTitle ?? 'کنسل',
-                    style: widget.cancelTextStyle ??
-                        TextStyle(color: widget.primaryColor),
-                  ),
+                8.0.spaceX,
+                CancelButton(
+                  buttonStyle: widget.cancelButtonStyle,
+                  borderRadius: widget.borderRadius,
+                  title: widget.cancelTitle,
+                  textStyle: widget.cancelTextStyle,
+                  primaryColor: widget.primaryColor,
                 ),
               ],
             ),
@@ -567,7 +333,8 @@ class _DatePickerDialogBoxWithoutDayState
     selectedYear = selectedDate.year;
 
     /// update selected date in text
-    selectedDateInText = '${monthsMap[selectedMonths]} $selectedYear';
+    selectedDateInText =
+        '${AppConstants.monthsMap[selectedMonths]} $selectedYear';
   }
 }